Corrected folder structure
