+++ resources/Hexominoes_Images/25.png
