+++ resources/Hexominoes_Images/27.png
