+++ resources/Hexominoes_Images/7.png
