+++ resources/Hexominoes_Images/Hexominoes.png
